Changed init.js to index.js Changed exported init to start moved logic in binwww.js to index.js Modified package.json scripts to reflect changes Removed passing ENV when launching production server - this seems not to be used anywhere and probably not needed at all like the development version does
